requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7